Extract FeedStoreSpy into its own file
