+++ EssentialFeed/EssentialFeedTests/Feed Cache/CacheFeedUseCaseTests.swift
@@ -131,45 +131,6 @@ final class CacheFeedUseCaseTests: XCTestCase {
         
     }
     
-    private class FeedStoreSpy: FeedStore {
-        
-        enum ReceivedMessage: Equatable {
-            case deleteCachedFeed
-            case insert(_ feed: [LocalFeedImage], timestamp: Date)
-        }
-        
-        private(set) var receivedMessages = [ReceivedMessage]()
-        
-        private var deletionCompletions = [Deletioncompletion]()
-        private var insertionCompletions = [Insertioncompletion]()
-        
-        func deleteCachedFeed(completion: @escaping Deletioncompletion) {
-            receivedMessages.append(.deleteCachedFeed)
-            deletionCompletions.append(completion)
-        }
-        
-        func completeDeletion(with error: Error, at index: Int = 0) {
-            deletionCompletions[index](error)
-        }
-        
-        func completeDeletionSuccessfully(at index: Int = 0) {
-            deletionCompletions[index](nil)
-        }
-        
-        func insert(_ feed: [LocalFeedImage], with timestamp: Date, completion: @escaping Insertioncompletion) {
-            receivedMessages.append(.insert(feed, timestamp: timestamp))
-            insertionCompletions.append(completion)
-        }
-        
-        func completeInsertion(with error: Error, at index: Int = 0) {
-            insertionCompletions[index](error)
-        }
-        
-        func completeInsertionSuccessfully(at index: Int = 0) {
-            insertionCompletions[index](nil)
-        }
-    }
-    
     private func uniqueImage() -> FeedImage {
         return FeedImage(id: UUID(), description: "", location: "", url: anyURL())
     }
